fmt/strtime: relax formatting of `%y` and `%g`
Previously, these would fail if the year wasn't in the range
`1969-2068`. This was done because parsing fails for the same reason,
and it seemed like good sense to have parsing and formatting be
consistent with one another.

... however, this is the *only* formatting specifier that can yield an
error based on the specific values provided. In contrast, all other
errors are tied to either the validity of the formatting string or if
all relevant data is provided (i.e., you get an error if you use `%y`
and your `BrokenDownTime` doesn't have a way to get the year).

I don't think this aberration is worth it. I think it's better to be
able to say the two broad error categories and be able to stick to them.
This should also hopefully avoid accidental panics, since now the only
error conditions that result in, say, `Zoned::strftime` panicking are
limited to the format string itself.

Fixes #428