fmt/strtime: make behavior of `BrokenDownTime::to_timestamp` a bit more consistent
Specifically, all of the other higher level constructors on
`BrokenDownTime` already try their "hardest" to produce the
requested value. Notably, both `BrokenDownTime::to_date` and
`BrokenDownTime::to_time` go to some effort to achieve this.

However, `BrokenDownTime::to_timestamp` wasn't doing everything it
could. Namely, it was completely ignoring its internal `timestamp`
field. This is weird and inconsistent from how the rest of the
constructors worked.

I think this is overall more flexible as well. Because it means that
parsing `%s` is, at the level of the individual pieces of data inside a
`BrokenDownTime`, completely separate from everything else. So you don't
have any weird overwriting behavior at the lowest levels.

I started thinking about this in #428. While I don't think the
motivation in #428 is that compelling, it got me thinking more deeply
about how a `BrokenDownTime` deals with timestamps. I think we got to
this state because I bolted `%s` on to `fmt::strtime` and didn't think
deeply about it.

Fixes #428