fmt/strtime: make `strptime` with `Zoned` and only `%s` work
Previously, if `%s` was parsed and there was no offset (`%z`) or IANA
time zone identifier (`%Q`), then trying to pull a `Zoned` out of a
`BrokenDownTime` would fail. But arguably, this should work. Namely, the
`strptime` APIs are already very relaxed about returning a `Zoned` with
fixed offset time zones (where the normal `FromStr` APIs forbid it due
to footguns). And since we have a way of representing an unknown time
zone, we can do exactly that.