fmt/friendly: simplify `set_span_unit_value`
This is a slight refactor to reduce a little duplication.

I don't mind duplication, but in this case, I was wondering why
this block was repeated and had to very carefully scrutinize the
code to follow it and make sure there weren't any differences. DRY
in this case I think makes it easier to understand.