fmt: slightly refactor duration parsing
This brings the code for parsing duration strings (ISO 8601 and
friendly) into `SignedDuration` or `Span` into more alignment.

There are still some annoying differences. And some bugs I
(re)discovered. They are marked as TODOs in this commit.

The main motivation here is to support parsing into
`std::time::Duration`. A third copy of this code was too unappealing to
me, so I'm trying to refactor the code to make it tighter and less
redundant.