signed_duration: add `i128` constructors
... and also a smattering of internal helpers.

The `SignedDuration::from_nanos_i128` constructor in particular
is analogous to `std::time::Duration::from_nanos_u128` coming to
a `std` near you. The rest are just a logical continuation.

This also adds some "get units with a remainder" methods that
I think will be useful in simplifying some of the duration
parsing code.