fmt: add support for a `&mut Vec<u8>` fast path
The printers can use this method to write directly into the
`Vec<u8>`'s spare capacity, instead of needing to write to our
new uninitialized buffer and then copy the data via the
`jiff::fmt::Write` interface.

Using the uninitialized buffer without this is still a dramatic
improvement. But this helps a bit more with minimal impact on
code size.