fmt: refactor RFC 2822 printer to use new uninitialized buffer abstraction
This uses the new code added in the previous two commits.

There should be no behavior changes here. We're just changing
the implementation to write directly into uninitialized data.
Whether it's a fixed size buffer on the stack or directly into
the spare capacity of a `String` or a `Vec<u8>`.

Note that this also adds a new error case to the RFC 2822 printer: when
rounding the offset would result in an out-of-bounds offset, we now
return an error. Previously, we would print an offset that Jiff would
then later fail to parse.