fmt: add more formatting routines to `BorrowedBuffer`
We'll need these to port more of Jiff's printers to the uninitialized
buffer abstraction.

I'm somewhat unhappy with the extra code here, but all of these routines
are intended to be inlined anyway. So this might end up saving code in
the long run.

I'm also somewhat unhappy with the fractional formatter. In particular,
the two loops in the case of an explicitly set precision. But I think an
explicitly set precision is the uncommon case. Oh well.

My loose plan is that we should ultimately be able to get rid of the
existing integer/fractional formatters. But we'll see.