fmt: refactor friendly printer to use padding/precision more explicitly
Basically, instead of creating and passing around an `IntegerFormatter`
or `FractionalFormatter`, we pass around padding and precision options
explicitly. We then materialize the formatters as needed. While this
looks odd, this is meant to prepare the code for a transition to
`BorrowedBuffer`, where we want the padding/precision settings at the
call site.

Probably the formatters should have just been simple functions from the
beginning. But they started off with more options before I was clear on
what the specific requirements were.