fmt: tweak `BorrowedBuffer::with_writer` API
Previously, we assumed that the size of the stack allocated buffer and
the runtime allocated extra capacity would always be the same. This is
okay for the case where there is very little difference between the
minimum and maximum lengths. But for RFC 9557 zoned datetimes, the
variance is quite high. Yet, we should also be able to compute the
capacity needed at runtime cheaply.

Thus, we change this API to accept possibly different values for the
uninitialized stack buffer and the runtime heap allocation.